Move bytes tests to tests/types/test_bytes.py
Also split the bytes cases out of the `test_default_validators`
parameterization and replace `conbytes()` usages with the explicit
annotated forms.
